fix: relocate filter segment buttons on org member list
