+++ apps/web/modules/members/members-view.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { DataTableProvider } from "@calcom/features/data-table/DataTableProvider";
 import MembersView from "@calcom/features/ee/organizations/pages/members";
 import Shell from "@calcom/features/shell/Shell";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
@@ -7,14 +8,23 @@ import { useLocale } from "@calcom/lib/hooks/useLocale";
 const MembersPage: React.FC = () => {
   const { t } = useLocale();
   return (
-    <Shell
-      withoutMain={false}
-      title={t("organization_members")}
-      description={t("organization_description")}
-      withoutSeo
-      subtitle={t("organization_description")}>
-      <MembersView />
-    </Shell>
+    <DataTableProvider defaultPageSize={25}>
+      <Shell
+        withoutMain={false}
+        title={t("organization_members")}
+        description={t("organization_description")}
+        withoutSeo
+        heading={t("organization_members")}
+        subtitle={t("organization_description")}
+        CTA={
+          <>
+            <p>hello</p>
+            <p>world</p>
+          </>
+        }>
+        <MembersView />
+      </Shell>
+    </DataTableProvider>
   );
 };
