+++ packages/features/data-table/components/filters/ColumnVisibilityButton.tsx
@@ -28,7 +28,7 @@ function ColumnVisibilityButtonComponent<TData>(
   {
     children,
     color = "secondary",
-    EndIcon = "sliders-vertical",
+    StartIcon = "sliders-vertical",
     table,
     ...rest
   }: ColumnVisiblityProps<TData> & ButtonProps,
@@ -41,8 +41,8 @@ function ColumnVisibilityButtonComponent<TData>(
   return (
     <Popover open={open} onOpenChange={setOpen}>
       <PopoverTrigger asChild>
-        <Button ref={ref} color={color} EndIcon={EndIcon} {...rest}>
-          {children ? children : t("View")}
+        <Button ref={ref} color={color} StartIcon={StartIcon} {...rest}>
+          {children ? children : t("display")}
         </Button>
       </PopoverTrigger>
       <PopoverContent className="w-[200px] p-0" align="start">
