+++ packages/features/data-table/index.ts
@@ -1,6 +1,6 @@
 export * from "./components";
 export * from "./lib/types";
 export * from "./lib/utils";
-export * from "./lib/context";
+export * from "./DataTableProvider";
 export * from "./lib/resizing";
 export * from "./hooks";
